+++ ogma/src/rt/mod.rs
@@ -4,4 +4,4 @@ pub mod bat;
 pub(crate) mod fscache;
 mod process;
 
-pub use process::process_expression;
+pub use process::{handle_help, process_expression};
