+++ ogma/src/rt/process.rs
@@ -23,7 +23,7 @@ where
     rt::fscache::ensure_init(root); // initialise the cache
 
     let expr = lang::syntax::parse::expression(expr, loc, defs).map_err(|e| e.0)?;
-    eng::handle_help(&expr, defs)?;
+    handle_help(&expr, defs)?;
     let vars = eng::Locals::default();
     let evaluator = eng::Evaluator::construct(I::as_type(), expr, defs, vars.clone())?;
     let cx = eng::Context {
@@ -34,4 +34,22 @@ where
     let output = evaluator.eval(seed.into(), cx)?.0;
 
     Ok(output)
+}
+
+/// Check if an expression has a help flag and output the help message (as the `Err` variant).
+pub fn handle_help(expr: &ast::Expression, definitions: &Definitions) -> Result<()> {
+    if let Some(block) = expr.blocks.get(0) {
+        let help = definitions.impls().get_help(&block.op())?;
+        if block
+            .terms()
+            .iter()
+            .any(|x| matches!(x, ast::Term::Flag(f) if f.str() == "help"))
+        {
+            Err(err::help_as_error(help))
+        } else {
+            Ok(())
+        }
+    } else {
+        Ok(())
+    }
 }
